mcp-server/requirements.txt
# MCP Server dependencies
# Install with: pip install -r mcp-server/requirements.txt

# FastAPI — the web framework for the MCP server and AI agent backend
fastapi==0.109.0

# Uvicorn — the ASGI server that runs FastAPI
uvicorn==0.27.0

# Pydantic — data validation for request/response models
pydantic==2.5.3

# Boto3 — AWS SDK for Python — used to read Terraform state from S3
boto3==1.34.0

# Requests — HTTP client for calling the GitHub Actions API
requests==2.31.0

# Anthropic — official Python SDK for the Claude API
anthropic==0.18.1

# python-dotenv — loads .env file into environment variables
python-dotenv==1.0.0
